requirements.txt
torch==2.5
transformers==4.48
datasets==3.0.1
pyyaml==6.0.2
accelerate==0.34.2